- Move all_children from beta to GA
