+++ lib/easypost/services.rb
@@ -10,7 +10,6 @@ require_relative 'services/api_key'
 require_relative 'services/batch'
 require_relative 'services/beta_rate'
 require_relative 'services/beta_referral_customer'
-require_relative 'services/beta_user'
 require_relative 'services/billing'
 require_relative 'services/carrier_account'
 require_relative 'services/carrier_metadata'
