--- spec/cassettes/beta_user/EasyPost_Services_BetaUser_get_next_page_retrieves_the_next_page_of_a_collection.yml
@@ -1,71 +1,0 @@
----
-http_interactions:
-- request:
-    method: get
-    uri: https://api.easypost.com/beta/users/children
-    body:
-      encoding: UTF-8
-      string: '{"page_size":5}'
-    headers:
-      Accept-Encoding:
-      - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
-      Accept:
-      - "*/*"
-      User-Agent: "<REDACTED>"
-      Host:
-      - api.easypost.com
-      Content-Type:
-      - application/json
-      Authorization: "<REDACTED>"
-  response:
-    status:
-      code: 200
-      message: OK
-    headers:
-      X-Frame-Options:
-      - SAMEORIGIN
-      X-Xss-Protection:
-      - 1; mode=block
-      X-Content-Type-Options:
-      - nosniff
-      X-Download-Options:
-      - noopen
-      X-Permitted-Cross-Domain-Policies:
-      - none
-      Referrer-Policy:
-      - strict-origin-when-cross-origin
-      X-Ep-Request-Uuid:
-      - 22687dad657ca103e78976a9000b95ac
-      Cache-Control:
-      - private, no-cache, no-store
-      Pragma:
-      - no-cache
-      Expires:
-      - '0'
-      Content-Type:
-      - application/json; charset=utf-8
-      X-Runtime:
-      - '0.049732'
-      Vary:
-      - Origin
-      Transfer-Encoding:
-      - chunked
-      X-Node:
-      - bigweb39nuq
-      X-Version-Label:
-      - easypost-202312142151-39776a9e02-master
-      X-Backend:
-      - easypost
-      X-Proxied:
-      - extlb1nuq 003ad9bca0
-      - intlb2nuq b3de2c47ef
-      Strict-Transport-Security:
-      - max-age=31536000; includeSubDomains; preload
-    body:
-      encoding: UTF-8
-      string: '{"children":[{"id":"user_484dd58db70a4f31b4bb862998cf0e04","object":"User","parent_id":"user_0f6b83e3530b401cb1e8aeaa6a250d4d","name":"Test
-        User","phone_number":"<REDACTED>","verified":true,"created_at":"2023-05-16T22:01:20Z"},{"id":"user_14e894c0d541459395f4456e7cf4f175","object":"User","parent_id":"user_0f6b83e3530b401cb1e8aeaa6a250d4d","name":"Test
-        User","phone_number":"<REDACTED>","verified":true,"created_at":"2023-09-27T22:05:26Z"},{"id":"user_f04df3dad13848339a7975d295d6629f","object":"User","parent_id":"user_0f6b83e3530b401cb1e8aeaa6a250d4d","name":"Test
-        User","phone_number":"<REDACTED>","verified":true,"created_at":"2023-11-30T19:23:22Z"}],"has_more":false}'
-  recorded_at: Fri, 15 Dec 2023 18:54:59 GMT
-recorded_with: VCR 6.1.0
